Move utils code into brick
